+++ include/core/globals.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "core/globals.hpp"
+#include "logger.hpp"
+
 class DirkEngine;
 
 extern DirkEngine* GEngine;
